JBRULES-537 BigDecimalFactory and BigIntegerFactory don't compile with java 1.5 source
git-svn-id: https://svn.jboss.org/repos/labs/labs/jbossrules/trunk@7342 c60d74c8-e8f6-0310-9e8f-d4a2fc68ab70

+++ drools-core/src/main/java/org/drools/base/evaluators/BigIntegerFactory.java
@@ -194,19 +194,19 @@ public class BigIntegerFactory
                                 final Object object1,
                                 final FieldValue object2) {
             final BigInteger comp = (BigInteger) extractor.getValue( object1 );
-            return comp.compareTo( object2.getValue() ) < 0;
+            return comp.compareTo( ( BigInteger ) object2.getValue() ) < 0;
         }
 
         public boolean evaluateCachedRight(final VariableContextEntry context,
                                            final Object left) {
             final BigInteger comp = (BigInteger) ((ObjectVariableContextEntry) context).right;
-            return comp.compareTo( context.declaration.getExtractor().getValue( left ) ) < 0;
+            return comp.compareTo( ( BigInteger ) context.declaration.getExtractor().getValue( left ) ) < 0;
         }
 
         public boolean evaluateCachedLeft(final VariableContextEntry context,
                                           final Object right) {
             final BigInteger comp = (BigInteger) context.extractor.getValue( right );
-            return comp.compareTo( ((ObjectVariableContextEntry) context).left ) < 0;
+            return comp.compareTo( ( BigInteger ) ((ObjectVariableContextEntry) context).left ) < 0;
         }
 
         public boolean evaluate(Extractor extractor1,
@@ -214,7 +214,7 @@ public class BigIntegerFactory
                                 Extractor extractor2,
                                 Object object2) {
             final BigInteger comp = (BigInteger) extractor1.getValue( object1 );
-            return comp.compareTo( extractor2.getValue( object2 ) ) < 0;
+            return comp.compareTo( ( BigInteger ) extractor2.getValue( object2 ) ) < 0;
         }
         
         public String toString() {
@@ -238,19 +238,19 @@ public class BigIntegerFactory
                                 final Object object1,
                                 final FieldValue object2) {
             final BigInteger comp = (BigInteger) extractor.getValue( object1 );
-            return comp.compareTo( object2.getValue() ) <= 0;
+            return comp.compareTo( ( BigInteger ) object2.getValue() ) <= 0;
         }
 
         public boolean evaluateCachedRight(final VariableContextEntry context,
                                            final Object left) {
             final BigInteger comp = (BigInteger) ((ObjectVariableContextEntry) context).right;
-            return comp.compareTo( context.declaration.getExtractor().getValue( left ) ) <= 0;
+            return comp.compareTo( ( BigInteger ) context.declaration.getExtractor().getValue( left ) ) <= 0;
         }
 
         public boolean evaluateCachedLeft(final VariableContextEntry context,
                                           final Object right) {
             final BigInteger comp = (BigInteger) context.extractor.getValue( right );
-            return comp.compareTo( ((ObjectVariableContextEntry) context).left ) <= 0;
+            return comp.compareTo( ( BigInteger ) ((ObjectVariableContextEntry) context).left ) <= 0;
         }
 
         public boolean evaluate(Extractor extractor1,
@@ -258,7 +258,7 @@ public class BigIntegerFactory
                                 Extractor extractor2,
                                 Object object2) {
             final BigInteger comp = (BigInteger) extractor1.getValue( object1 );
-            return comp.compareTo( extractor2.getValue( object2 ) ) <= 0;
+            return comp.compareTo( ( BigInteger ) extractor2.getValue( object2 ) ) <= 0;
         }
         
         public String toString() {
@@ -282,19 +282,19 @@ public class BigIntegerFactory
                                 final Object object1,
                                 final FieldValue object2) {
             final BigInteger comp = (BigInteger) extractor.getValue( object1 );
-            return comp.compareTo( object2.getValue() ) > 0;
+            return comp.compareTo( ( BigInteger ) object2.getValue() ) > 0;
         }
 
         public boolean evaluateCachedRight(final VariableContextEntry context,
                                            final Object left) {
             final BigInteger comp = (BigInteger) ((ObjectVariableContextEntry) context).right;
-            return comp.compareTo( context.declaration.getExtractor().getValue( left ) ) > 0;
+            return comp.compareTo( ( BigInteger ) context.declaration.getExtractor().getValue( left ) ) > 0;
         }
 
         public boolean evaluateCachedLeft(final VariableContextEntry context,
                                           final Object right) {
             final BigInteger comp = (BigInteger) context.extractor.getValue( right );
-            return comp.compareTo( ((ObjectVariableContextEntry) context).left ) > 0;
+            return comp.compareTo( ( BigInteger ) ((ObjectVariableContextEntry) context).left ) > 0;
         }
 
         public boolean evaluate(Extractor extractor1,
@@ -302,7 +302,7 @@ public class BigIntegerFactory
                                 Extractor extractor2,
                                 Object object2) {
             final BigInteger comp = (BigInteger) extractor1.getValue( object1 );
-            return comp.compareTo( extractor2.getValue( object2 ) ) > 0;
+            return comp.compareTo( ( BigInteger ) extractor2.getValue( object2 ) ) > 0;
         }
         
         public String toString() {
@@ -326,19 +326,19 @@ public class BigIntegerFactory
                                 final Object object1,
                                 final FieldValue object2) {
             final BigInteger comp = (BigInteger) extractor.getValue( object1 );
-            return comp.compareTo( object2.getValue() ) >= 0;
+            return comp.compareTo( ( BigInteger ) object2.getValue() ) >= 0;
         }
 
         public boolean evaluateCachedRight(final VariableContextEntry context,
                                            final Object left) {
             final BigInteger comp = (BigInteger) ((ObjectVariableContextEntry) context).right;
-            return comp.compareTo( context.declaration.getExtractor().getValue( left ) ) >= 0;
+            return comp.compareTo( ( BigInteger ) context.declaration.getExtractor().getValue( left ) ) >= 0;
         }
 
         public boolean evaluateCachedLeft(final VariableContextEntry context,
                                           final Object right) {
             final BigInteger comp = (BigInteger) context.extractor.getValue( right );
-            return comp.compareTo( ((ObjectVariableContextEntry) context).left ) >= 0;
+            return comp.compareTo( ( BigInteger ) ((ObjectVariableContextEntry) context).left ) >= 0;
         }
 
         public boolean evaluate(Extractor extractor1,
@@ -346,7 +346,7 @@ public class BigIntegerFactory
                                 Extractor extractor2,
                                 Object object2) {
             final BigInteger comp = (BigInteger) extractor1.getValue( object1 );
-            return comp.compareTo( extractor2.getValue( object2 ) ) >= 0;
+            return comp.compareTo( ( BigInteger ) extractor2.getValue( object2 ) ) >= 0;
         }
         
         public String toString() {
